+++ app/src/main/java/me/tiptap/tiptap/main/MainViewPagerAdapter.kt
@@ -1,13 +1,12 @@
-package me.tiptap.tiptap.adapter
+package me.tiptap.tiptap.main
 
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentManager
 import android.support.v4.app.FragmentStatePagerAdapter
 import me.tiptap.tiptap.diaries.DiariesFragment
-import me.tiptap.tiptap.main.MainFragment
 import me.tiptap.tiptap.scratch.ScratchFragment
 
-class MainViewPagerAdapter(fm: FragmentManager, val pageCount: Int) : FragmentStatePagerAdapter(fm) {
+class MainViewPagerAdapter(fm: FragmentManager, private val pageCount: Int) : FragmentStatePagerAdapter(fm) {
 
 
     override fun getItem(position: Int): Fragment = when (position) {
